Restore direct use of the input text for Markdown `Text`
PR #24388 changed the markdown parsing to copy parsed text in order
to handle markdown escaping, removing the optimization to instead
reuse text from the input.

Another issue with that change was that handling of finding links
within `Text` intermixed use of `text` and `parsed`, relying on the
offsets matching up (which I believe was true in practice).

The solution is to distinguish pulldown_cmark `Text` nodes that share
bytes with the input and those that do not. Text nodes that do not
share bytes with the input are concatenated together into a
`SharedString`.